Update locker rental information (#236)
This PR updates the locker rental information, and moves the page into a
separate card in the `/services/` route. I replaced the Coffee Chats
card with this one, as we're not currently running Coffee Chats.

I also added a new image for the lockers - the images have to all be the
same size, and I didn't want to crop the existing SVG too tightly.