Fix setInterval → setTimeout (#257)
Currently, the newsletter signup modal will continue to spawn the
darkened background at an interval of 10 seconds. If a user navigates to
our page for the first time, navigates off the page for a couple
minutes, then comes back and clicks on one of the buttons, the
backgrounds won't disappear, so they'll see a completely dark page.

Should be fixed with `setTimeout`, which runs only once.