Add shortcode for clubs unified calendar (#261)
This PR is part of #255 - it adds an events calendar similar to the one
on our homepage, but can include multiple event feeds. The shortcode
takes in any number of hosted ICS files (calendar event format files
that can be exposed through GCal), and displays them on the clubs page.

Currently, only the CSSS events are shown, but other feeds can easily be
added pending more outreach for club events!

Let me know if an exec's email should be listed instead of redirecting
to the suggestions box.