New review for CPSC 221 by Arman (#321)
> Enjoyed the course and the way it was structured. Content is pretty
interesting and not too bad to learn, but the assignments can be
challenging. All in all good course with goated profs.
>
> Difficulty: 4/5
> Quality: 5/5
> <cite><a href='https://github.com/armanmoztar'>Arman</a>, Nov 25 2022,
course taken during 2021W2</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Arman
    authorLink: https://github.com/armanmoztar
    date: 2022-11-25
    review: |
Enjoyed the course and the way it was structured. Content is pretty
interesting and not too bad to learn, but the assignments can be
challenging. All in all good course with goated profs.
    difficulty: 4
    quality: 5
    sessionTaken: 2021W2

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker