Add new batch of companies (#340)
This PR adds a new batch of companies as of 5 January 2023 per the
spreadsheet (see Slack).

Several companies (CSIS, West Fraser Timber) did not provide logos, so
I've found those online and will replace them in the next batch if
different logos are provided. Several companies also have not been
confirmed yet in Column E, and will be added in the next batch when
confirmed.