New review for CPSC 304 by annonn (#371)
> Teaches all the basic you need to get started with Relational
Databases. Highly recommend taking this course even though it is not a
required course for the CPSC major. The project was easy and almost
everyone got a 100% in it. Course final and assignments were also easy.
Not too heavy course load.
>
> Difficulty: 2.5/5
> Quality: 4/5
> <cite><a href=''>annonn</a>, Feb 03 2023, course taken during
2022W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: annonn
    authorLink: 
    date: 2023-02-03
    review: |
Teaches all the basic you need to get started with Relational Databases.
Highly recommend taking this course even though it is not a required
course for the CPSC major. The project was easy and almost everyone got
a 100% in it. Course final and assignments were also easy. Not too heavy
course load.
    difficulty: 2.5
    quality: 4
    sessionTaken: 2022W1

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker