New review for CPSC 404 by HK (#379)
> Course is taught well but the material becomes complicated pretty
quickly. Took it w Ed Knorr and he had a student-friendly grading policy
(3 midterms with worst one being replaced by your final if the latter is
better). Each midterm was a month apart and no major assignments that
take over your weekends. You could get 13% right-away on the course via
clickers, In-Class/Pre-class exercises that are graded based partially
on effort and correctness.


**Note for the dev team:
This is like a (awesome) curated version of RMP! Would be useful to see
which profs teach a given course and filter by prof to see how
ratings/difficulties change!
>
> Difficulty: 4.5/5
> Quality: 5/5
> <cite><a href=''>HK</a>, Feb 06 2023, course taken during
2022W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: HK
    authorLink: 
    date: 2023-02-06
    review: |
Course is taught well but the material becomes complicated pretty
quickly. Took it w Ed Knorr and he had a student-friendly grading policy
(3 midterms with worst one being replaced by your final if the latter is
better). Each midterm was a month apart and no major assignments that
take over your weekends. You could get 13% right-away on the course via
clickers, In-Class/Pre-class exercises that are graded based partially
on effort and correctness.
      
      
      **Note for the dev team:
This is like a (awesome) curated version of RMP! Would be useful to see
which profs teach a given course and filter by prof to see how
ratings/difficulties change!
    difficulty: 4.5
    quality: 5
    sessionTaken: 2022W1

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker

---------

Co-authored-by: Asad Dhorajiwala <asadrehandhorajiwala@gmail.com>