New review for CPSC 314 by KF (#394)
> Took this course with Dinesh Pai. If you've done matrix algebra,
you'll be fine in terms of the math requirement for this course.
Assignments were interesting but could be time-consuming in debugging.
Good if you like JavaScript. Lectures focused on concepts behind
computer graphics, shaders, etc, but did not help as much with the
coding portion. Had to look online for learning a lot of Three.js.
Almost half the midterm questions were the same as the ones in the
quizzes / practice exams. Questions were also reused in final. Overall
interesting course.
>
> Difficulty: 3.5/5
> Quality: 3.5/5
> <cite><a href=''>KF</a>, Feb 15 2023, course taken during
2021W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: KF
    authorLink: 
    date: 2023-02-15
    review: |
Took this course with Dinesh Pai. If you've done matrix algebra, you'll
be fine in terms of the math requirement for this course. Assignments
were interesting but could be time-consuming in debugging. Good if you
like JavaScript. Lectures focused on concepts behind computer graphics,
shaders, etc, but did not help as much with the coding portion. Had to
look online for learning a lot of Three.js. Almost half the midterm
questions were the same as the ones in the quizzes / practice exams.
Questions were also reused in final. Overall interesting course.
    difficulty: 3.5
    quality: 3.5
    sessionTaken: 2021W1

<pre>
</details>
This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker