New review for CPSC 320 by Kristen F (#393)
> Took this with Patrice Belleville. Extremely difficult course, but the
course was ran well. Assignments were very tough, I needed to go to
office hours and consult with peers very often for understanding of
concepts. Prof was understanding and fair. Start assignments early and
go to lectures & office hours, assignments can take up a significant
portion of time to understand and office hours are often quite packed.
Concepts taught in the course are very useful for industry and leetcode.
>
> Difficulty: 5/5
> Quality: 4.5/5
> <cite><a href=''>Kristen F</a>, Feb 15 2023, course taken during
2022W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Kristen F
    authorLink: 
    date: 2023-02-15
    review: |
Took this with Patrice Belleville. Extremely difficult course, but the
course was ran well. Assignments were very tough, I needed to go to
office hours and consult with peers very often for understanding of
concepts. Prof was understanding and fair. Start assignments early and
go to lectures & office hours, assignments can take up a significant
portion of time to understand and office hours are often quite packed.
Concepts taught in the course are very useful for industry and leetcode.
    difficulty: 5
    quality: 4.5
    sessionTaken: 2022W1

<pre>
</details>
This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker