New review for CPSC 310 by Kristen F (#392)
> Course was very much split into lecture content and project content.
Lecture content was focused on Agile concepts and general industry terms
and ideas, and was not difficult. Project had a quite intensive
workload, but the end result of a full stack project is quite good on a
resume. Personally found that the project coding was not extremely
difficult, just time-consuming, and also required learning Promises.
>
> Difficulty: 4/5
> Quality: 4.5/5
> <cite><a href=''>Kristen F</a>, Feb 15 2023, course taken during
2020W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Kristen F
    authorLink: 
    date: 2023-02-15
    review: |
Course was very much split into lecture content and project content.
Lecture content was focused on Agile concepts and general industry terms
and ideas, and was not difficult. Project had a quite intensive
workload, but the end result of a full stack project is quite good on a
resume. Personally found that the project coding was not extremely
difficult, just time-consuming, and also required learning Promises.
    difficulty: 4
    quality: 4.5
    sessionTaken: 2020W1

<pre>
</details>
This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker