New review for CPSC 430 by Kristen F (#391)
> An ethics course focused on discussion about ethical problems within
the tech sector. Took the course with Kevin Layton-Brown and Giulia
Toti. Each week, there is a paragraph that you need to write
(approximately 1500 characters (not words)), which will be peer-graded
later. Depending on your previous amount of weekly work, you peer-grade
3 or more of your fellow peers' paragraphs (anonymously). No programming
involved, but a lot of writing. Something to keep in mind if you find
your writing skills are not as strong.
>
> Difficulty: 2/5
> Quality: 4/5
> <cite><a href=''>Kristen F</a>, Feb 15 2023, course taken during
2022W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Kristen F
    authorLink: 
    date: 2023-02-15
    review: |
An ethics course focused on discussion about ethical problems within the
tech sector. Took the course with Kevin Layton-Brown and Giulia Toti.
Each week, there is a paragraph that you need to write (approximately
1500 characters (not words)), which will be peer-graded later. Depending
on your previous amount of weekly work, you peer-grade 3 or more of your
fellow peers' paragraphs (anonymously). No programming involved, but a
lot of writing. Something to keep in mind if you find your writing
skills are not as strong.
    difficulty: 2
    quality: 4
    sessionTaken: 2022W1

<pre>
</details>
This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker