New review for CPSC 404 by KF (#390)
> Course is well-taught and not too difficult, but quickly gains a lot
of content and can be pretty dry at times. No programming or very little
involved (some SQL). Took this class with Ed Knorr, his marking scheme
is very helpful for students. There are pre-class and in-class exercises
with frequent deadlines, but they're not difficult. Assignments had good
instructions on what to follow.
>
> Difficulty: 3/5
> Quality: 5/5
> <cite><a href=''>KF</a>, Feb 14 2023, course taken during
2022W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: KF
    authorLink: 
    date: 2023-02-14
    review: |
Course is well-taught and not too difficult, but quickly gains a lot of
content and can be pretty dry at times. No programming or very little
involved (some SQL). Took this class with Ed Knorr, his marking scheme
is very helpful for students. There are pre-class and in-class exercises
with frequent deadlines, but they're not difficult. Assignments had good
instructions on what to follow.
    difficulty: 3
    quality: 5
    sessionTaken: 2022W1

<pre>
</details>
This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker