Fix Prettier workflow (#395)
This PR closes #356 and fixes the Prettier workflow to correctly run on
push to `master`.

Previously, the workflow couldn't run because its default authorization
token didn't have permissions to bypass the branch protection filter
currently set on `master`. This has been fixed by generating a personal
access token for the `csssbot` account, granting it admin permissions
for this repository, and passing in the token to the Prettier action.

I've also added a `workflow_dispatch` trigger so we can run the workflow
manually if needed.