New review for CPSC 313 by FC (#403)
> Great courses specially you are taking it with Margo. Course content
is managed well on PrairieLearn. A lot of programming assignments
especially the second half of the course.

Systems can be messy but you will learn a lot for sure
>
> Difficulty: 3/5
> Quality: 5/5
> <cite><a href=''>FC</a>, Feb 20 2023, course taken during
2020W2</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: FC
    authorLink: 
    date: 2023-02-20
    review: |
Great courses specially you are taking it with Margo. Course content is
managed well on PrairieLearn. A lot of programming assignments
especially the second half of the course.
      
      Systems can be messy but you will learn a lot for sure
    difficulty: 3
    quality: 5
    sessionTaken: 2020W2

<pre>
</details>
<details><summary>View raffle email</summary>
ykarry@student.ubc.ca
Remember to edit this out (… > 'Edit') after processing the review.
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker