New review for CPSC 340 by ayylmao (#405)
> Even for a CS course the material is extremely dense and there is a
great deal of math involved. But the instructors are excellent and you
definitely learn a ton.

Course comprises of bi-weekly assignments. Prepare to spend 20+ hours on
the assignment if you do it individually. First day of class we got an
assignment that was 20 pages long, each page with multiple questions
and/or a small coding portion involved. Highly recommend getting a good
partner.
>
> Difficulty: 5/5
> Quality: 5/5
> <cite><a href=''>ayylmao</a>, Feb 22 2023, course taken during
2023W2</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: ayylmao
    authorLink: 
    date: 2023-02-22
    review: |
Even for a CS course the material is extremely dense and there is a
great deal of math involved. But the instructors are excellent and you
definitely learn a ton.
      
Course comprises of bi-weekly assignments. Prepare to spend 20+ hours on
the assignment if you do it individually. First day of class we got an
assignment that was 20 pages long, each page with multiple questions
and/or a small coding portion involved. Highly recommend getting a good
partner.
    difficulty: 5
    quality: 5
    sessionTaken: 2023W2

<pre>
</details>
<details><summary>View raffle email</summary>
ykim11@student.ubc.ca
Remember to edit this out (… > 'Edit') after processing the review.
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker