New review for CPSC 317 by ayylmao (#404)
> Like all CPSC courses, the material is quite dense and the assignments
are heavy. I took this course with Norm / Aastha and both are very
competent lecturers who clearly spend a lot of time helping students.
The piazza helpfulness rate is outstanding.

When I took this course it had 5 quizzes, 6% each, for a total of 30%,
instead of midterms. The quizzes are close-book, biweekly and proctored
(1 hour long). They do a good job of making you stay on top of the
material and I overall prefer the quizzes.

The assignments are biweekly and also worth around the same amount. But
like many CS assignments depends heavily on your programming skills. So
far I found the assignments to be challenging and at times vague (but
the piazza helps for clarifications). Prepare to spend 10-20 hours per
assignment if you are a decent programmer, more otherwise.

Material is difficult conceptually. There is just a lot of it. Very
relevant to real life. Recommend as an elective.
>
> Difficulty: 4/5
> Quality: 4/5
> <cite><a href=''>ayylmao</a>, Feb 22 2023, course taken during
2023W2</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: ayylmao
    authorLink: 
    date: 2023-02-22
    review: |
Like all CPSC courses, the material is quite dense and the assignments
are heavy. I took this course with Norm / Aastha and both are very
competent lecturers who clearly spend a lot of time helping students.
The piazza helpfulness rate is outstanding.
      
When I took this course it had 5 quizzes, 6% each, for a total of 30%,
instead of midterms. The quizzes are close-book, biweekly and proctored
(1 hour long). They do a good job of making you stay on top of the
material and I overall prefer the quizzes.
      
The assignments are biweekly and also worth around the same amount. But
like many CS assignments depends heavily on your programming skills. So
far I found the assignments to be challenging and at times vague (but
the piazza helps for clarifications). Prepare to spend 10-20 hours per
assignment if you are a decent programmer, more otherwise.
      
Material is difficult conceptually. There is just a lot of it. Very
relevant to real life. Recommend as an elective.
    difficulty: 4
    quality: 4
    sessionTaken: 2023W2

<pre>
</details>
<details><summary>View raffle email</summary>
ykim11@student.ubc.ca
Remember to edit this out (… > 'Edit') after processing the review.
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker