New review for CPSC 436 by Brandon Orange  (#409)
> With Vered, 436N was an excellent course. Slightly unstructured but
that is the nature of learning new models and the 436's. The course had
challenging assignments and a medium difficulty midterm/final. Highly
recommended. I also think it's better than 440.
>
> Difficulty: 4/5
> Quality: 4/5
> <cite><a href=''>Brandon Orange </a>, Feb 25 2023, course taken during
2022W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Brandon Orange 
    authorLink: 
    date: 2023-02-25
    review: |
With Vered, 436N was an excellent course. Slightly unstructured but that
is the nature of learning new models and the 436's. The course had
challenging assignments and a medium difficulty midterm/final. Highly
recommended. I also think it's better than 440.
    difficulty: 4
    quality: 4
    sessionTaken: 2022W1

<pre>
</details>
<details><summary>View raffle email</summary>
b.tong@student.ubc.ca
Remember to edit this out (… > 'Edit') after processing the review.
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker