Add year requirement (#411)
This PR fixes an issue I noticed in #408 - the course review year input
box is not `required`, so a user left it blank and their review date
only showed as `W1`. I've added a `required` attribute to fix this.

(Also, someone came to the Cube before reading break asking why their
course review wasn't published, so I've left an additional note in the
submission confirmation to allow for processing time. Happy to take
suggestions on how to make this even clearer!)