New review for CPSC 210 by Skyler (#435)
> I had Steven, his teaching was excellent, he is clearly and effective
in everything he teaches. The course is modularised very effectively. I
wish there was more skepticism on OOP and there is a heavy reliance on
the debugger and other IDE tools which is not great. The content is
engaging and the "fliiped" lecture style is really helpful, especially
since Steven will sit down next to you and help you through problems in
a really helpful way. He is also helpful if you ask questions that reach
outside the exact course content. I would reccomend coming in with some
experience in java or other (non-racket) languages, the concepts are
important.
>
> Difficulty: 3.5/5
> Quality: 5/5
> <cite><a href='https://github.students.cs.ubc.ca/ssauer02'>Skyler</a>,
Apr 05 2023, course taken during 2023W2</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Skyler
    authorLink: https://github.students.cs.ubc.ca/ssauer02
    date: 2023-04-05
    review: |
I had Steven, his teaching was excellent, he is clearly and effective in
everything he teaches. The course is modularised very effectively. I
wish there was more skepticism on OOP and there is a heavy reliance on
the debugger and other IDE tools which is not great. The content is
engaging and the "fliiped" lecture style is really helpful, especially
since Steven will sit down next to you and help you through problems in
a really helpful way. He is also helpful if you ask questions that reach
outside the exact course content. I would reccomend coming in with some
experience in java or other (non-racket) languages, the concepts are
important.
    difficulty: 3.5
    quality: 5
    sessionTaken: 2023W2

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker

Co-authored-by: Ramit Kataria <ramit.kataria@gmail.com>