New review for CPSC 320 by Crispin (#439)
> One of the more difficult course in CPSC as it delves into challenging
abstract topics. Assignments will take a large portion of your time.
Though challenging, it is an important course that can help leapfrog
your leetcode skills to be more competitive for internships.
>
> Difficulty: 5/5
> Quality: 4.5/5
> <cite><a href=''>Crispin</a>, Apr 11 2023, course taken during
2022W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Crispin
    authorLink: 
    date: 2023-04-11
    review: |
One of the more difficult course in CPSC as it delves into challenging
abstract topics. Assignments will take a large portion of your time.
Though challenging, it is an important course that can help leapfrog
your leetcode skills to be more competitive for internships.
    difficulty: 5
    quality: 4.5
    sessionTaken: 2022W1

<pre>
</details>
This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker