New review for CPSC 422 by 422student (#440)
> This course is a step up from 322 because we actually get experience
learning more up-to-date tools. But the system is lacking the in the
sense that courses like 340 teach you a lot more, a lot quicker. But if
you want a chill course take 422.
>
> Difficulty: 3.5/5
> Quality: 3.5/5
> <cite><a href=''>422student</a>, Apr 23 2023, course taken during
2021W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: 422student
    authorLink: 
    date: 2023-04-23
    review: |
This course is a step up from 322 because we actually get experience
learning more up-to-date tools. But the system is lacking the in the
sense that courses like 340 teach you a lot more, a lot quicker. But if
you want a chill course take 422.
    difficulty: 3.5
    quality: 3.5
    sessionTaken: 2021W1

<pre>
</details>
This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker