New review for CPSC 210 by ubcstudent2 (#443)
> The concepts in this course are useful. The course overall is good,
but I have a big issue with PraireLearn just makes everything so much
harder for no reason. In this class we use a system called PrairieLearn
for exams and practice exams. To put it frankly, this system is
horrible. First of all, why are we writing code in a web browser? This
course should learn from its pre-req CPSC 110 and move onto programming
in an IDE for exams and all practices. Also it should really have
autograders. Trying to match our answers with the solution when the
solution can be often coded in maybe ways is horrible.

For practice problems, you can only answer the question once, if you
want to try again, you have to creat another instance of the whole
practice set. The text editor is bulky and annoying. Exams seriously
tests more on predicting what the question writer wants. It is also very
hard to perfect grades, because of multiple choice. You can easily get 0
on a question if you just miss a few details.

PraireLearn randomizes the order of answer choices for some reason.
Consider this arbitrary case , if the question involves 3 variables,
usually the question would list the answer choices (1 1), (1 2), (2, 1),
(2, 2), (3, 1), (3, 2) or in some other order. With the order randomized
it could be (2, 2), (1 2), (1 1), (3, 1) , (2, 1), (3, 2). Which messes
up the flow of natural problem solving especially when the answer
choices are full sentences and you need to pick out which parts are
different.

PraireLearn randomizes variable names. Again, this makes every
unnecessarily harder for no reason. They love using arbitrary names with
no meaning at all like "smurf", "lorem", "ipsum." Again, you have to
look back and forth to the question. Even if they just used random
everyday words, it would make it easier.

The questions are just not that clear. If you have a slightly different
interpretation you can expect a maximum of 50% on that question.

Overall the course material is very very easy, but the PraireLearn
system just suck. 0/10. It adds so much extra resistance to the
otherwise very easy course material. CPSC 210 can easily be a 3 credit
course. Also its hard to wrap my head around how a "second year"
computer science course at "a good" university has so much overlap with
Programming 10, 11, 12 in highschools.

>
> Difficulty: 2/5
> Quality: 2/5
> <cite><a href=''>ubcstudent2</a>, Apr 27 2023, course taken during
2022W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: ubcstudent2
    authorLink: 
    date: 2023-04-27
    review: |
The concepts in this course are useful. The course overall is good, but
I have a big issue with PraireLearn just makes everything so much harder
for no reason. In this class we use a system called PrairieLearn for
exams and practice exams. To put it frankly, this system is horrible.
First of all, why are we writing code in a web browser? This course
should learn from its pre-req CPSC 110 and move onto programming in an
IDE for exams and all practices. Also it should really have autograders.
Trying to match our answers with the solution when the solution can be
often coded in maybe ways is horrible.
      
For practice problems, you can only answer the question once, if you
want to try again, you have to creat another instance of the whole
practice set. The text editor is bulky and annoying. Exams seriously
tests more on predicting what the question writer wants. It is also very
hard to perfect grades, because of multiple choice. You can easily get 0
on a question if you just miss a few details.
      
PraireLearn randomizes the order of answer choices for some reason.
Consider this arbitrary case , if the question involves 3 variables,
usually the question would list the answer choices (1 1), (1 2), (2, 1),
(2, 2), (3, 1), (3, 2) or in some other order. With the order randomized
it could be (2, 2), (1 2), (1 1), (3, 1) , (2, 1), (3, 2). Which messes
up the flow of natural problem solving especially when the answer
choices are full sentences and you need to pick out which parts are
different.
      
PraireLearn randomizes variable names. Again, this makes every
unnecessarily harder for no reason. They love using arbitrary names with
no meaning at all like "smurf", "lorem", "ipsum." Again, you have to
look back and forth to the question. Even if they just used random
everyday words, it would make it easier.
      
The questions are just not that clear. If you have a slightly different
interpretation you can expect a maximum of 50% on that question.
      
Overall the course material is very very easy, but the PraireLearn
system just suck. 0/10. It adds so much extra resistance to the
otherwise very easy course material. CPSC 210 can easily be a 3 credit
course. Also its hard to wrap my head around how a "second year"
computer science course at "a good" university has so much overlap with
Programming 10, 11, 12 in highschools.
      
    difficulty: 2
    quality: 2
    sessionTaken: 2022W1

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker