New review for CPSC 110 by Jen Chen (#467)
> I found the 100-level UBC CS Courses to be among the easiest in my
degree. All the content (tested, otherwise) was fair / expected. Nothing
too difficult as long as you keep up with the course-work.
>
> Difficulty: 1/5
> Quality: 5/5
> <cite><a href=''>Jen Chen</a>, Jul 20 2023, course taken during
2018W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: Jen Chen
    authorLink: 
    date: 2023-07-20
    review: |
I found the 100-level UBC CS Courses to be among the easiest in my
degree. All the content (tested, otherwise) was fair / expected. Nothing
too difficult as long as you keep up with the course-work.
    difficulty: 1
    quality: 5
    sessionTaken: 2018W1

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker