Update Food Price List (#533)
Someone came to the cube and noted that the price on the website is
different from the actually price of products. So I updated a few items
in Food Price List. I didn't do an in depth update because our new
system is under development. So I just updated some of the prices.