New review for CPSC 416 by michael (#546)
> I like the course, which uses a project-based approach to let the
students explore and learn the concepts we were taught in the lecture.
I'd advocate more STEM courses like this course to have a final project
or maybe case studies. The only downside is that the reading tasks are
heavy.
>
> Difficulty: 4/5
> Quality: 4.5/5
> <cite><a href=''>michael</a>, Dec 12 2023, course taken during
2023W1</cite>
<details><summary>View YAML for new review</summary>
<pre>
  - author: michael
    authorLink: 
    date: 2023-12-12
    review: |
I like the course, which uses a project-based approach to let the
students explore and learn the concepts we were taught in the lecture.
I'd advocate more STEM courses like this course to have a final project
or maybe case studies. The only downside is that the reading tasks are
heavy.
    difficulty: 4
    quality: 4.5
    sessionTaken: 2023W1

<pre>
</details>This is an auto-generated PR made using:
https://github.com/ubccsss/course-review-worker